postgres_to_es/requirements.txt
psycopg2-binary==2.9.6
elasticsearch==8.7.0
